requirements.txt
fastapi==0.115.12
uvicorn[standard]
pydantic>=2.7,<3
openai>=1.105.0
google-genai>=1.33.0
jsonschema>=4.23.0

